requirements.txt
alembic==1.0.7
Babel==2.6.0
Click==7.0
elasticsearch==7.0.0
Flask==1.0.2
Flask-Babel==0.12.2
Flask-Bootstrap4==4.0.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.3.1
Flask-Moment==0.7.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
guess-language-spirit==0.5.3
httplib2==0.9.2
Jinja2==2.10
language-selector==0.1
migrate==0.3.8
mysql-connector==2.1.6
pygobject==3.26.1
PyJWT==1.7.1
python-dotenv==0.10.1
requests==2.21.0
SQLAlchemy==1.2.17
system-service==0.3
systemd-python==234
urllib3==1.24.1
Werkzeug==0.15.0
WTForms==2.2.1

